Clean up some examples.
