+++ test/jdk/java/lang/reflect/code/TestLinqUsingQuoted.java
@@ -25,13 +25,12 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 
 import java.lang.reflect.code.Op;
-import java.lang.reflect.code.Quotable;
+import java.lang.reflect.code.Quoted;
+import java.lang.reflect.code.Value;
 import java.lang.reflect.code.descriptor.MethodDesc;
 import java.lang.reflect.code.descriptor.TypeDesc;
 import java.lang.reflect.code.interpreter.Interpreter;
 import java.lang.invoke.MethodHandles;
-import java.util.function.Function;
-import java.util.function.Predicate;
 import java.util.stream.Stream;
 
 import static java.lang.reflect.code.descriptor.MethodDesc.method;
@@ -41,23 +40,14 @@ import static java.lang.reflect.code.op.CoreOps.*;
 
 /*
  * @test
- * @run testng TestLinqUsingQuotable
+ * @run testng TestLinqUsingQuoted
  */
 
-public class TestLinqUsingQuotable {
-
-    // Quotable functional interfaces
-
-    public interface QuotablePredicate<T> extends Quotable, Predicate<T> {
-    }
-
-    public interface QuotableFunction<T, R> extends Quotable, Function<T, R> {
-    }
-
+public class TestLinqUsingQuoted {
 
     // Query interfaces
 
-    public interface Queryable<T> {
+    public interface Queryable {
         TypeDesc elementType();
 
         // Queryable<T> -> Queryable<U>
@@ -65,72 +55,80 @@ public class TestLinqUsingQuotable {
 
         QueryProvider provider();
 
-        @SuppressWarnings("unchecked")
-        default Queryable<T> where(QuotablePredicate<T> f) {
-            LambdaOp l = (LambdaOp) f.quoted().op();
-            return (Queryable<T>) insertQuery(elementType(), "where", l);
+        // T -> boolean
+        // Predicate<T>
+        default Queryable where(Quoted f) {
+            // @@@@ validate
+            ClosureOp c = (ClosureOp) f.op();
+            return insertQuery(elementType(), "where", c);
         }
 
-        @SuppressWarnings("unchecked")
-        default <R> Queryable<R> select(QuotableFunction<T, R> f) {
-            LambdaOp l = (LambdaOp) f.quoted().op();
-            return (Queryable<R>) insertQuery(l.funcDescriptor().returnType(), "select", l);
+        // T -> R
+        // Function<T, R>
+        default Queryable select(Quoted f) {
+            // @@@@ validate
+            ClosureOp c = (ClosureOp) f.op();
+            return insertQuery(c.funcDescriptor().returnType(), "select", c);
         }
 
-        private Queryable<?> insertQuery(TypeDesc elementType, String methodName, LambdaOp lambdaOp) {
+        private Queryable insertQuery(TypeDesc et, String name, ClosureOp c) {
             QueryProvider qp = provider();
 
-            // Copy function expression, replacing return operation
             FuncOp currentQueryExpression = expression();
-            TypeDesc queryableType = TypeDesc.type(qp.queryableType(), elementType);
-            FuncOp nextQueryExpression = func("query",
-                    methodType(queryableType, currentQueryExpression.funcDescriptor().parameters()))
-                    .body(b -> b.inline(currentQueryExpression, b.parameters(), (block, query) -> {
-                        Op.Result fi = block.op(lambdaOp);
+            FuncOp nextQueryExpression = currentQueryExpression.transform((block, op) -> {
+                if (op instanceof ReturnOp rop && rop.ancestorBody() == currentQueryExpression.body()) {
+                    Value query = block.context().getValue(rop.returnValue());
 
-                        MethodDesc md = method(qp.queryableType(), methodName,
-                                methodType(qp.queryableType(), lambdaOp.functionalInterface()).erase());
-                        Op.Result queryable = block.op(invoke(queryableType, md, query, fi));
+                    Op.Result quotedLambda = block.op(quoted(block.parentBody(), qblock -> c));
 
-                        block.op(_return(queryable));
-                    }));
+                    MethodDesc md = method(qp.queryableType(), name,
+                            methodType(qp.queryableType(), QuotedOp.QUOTED_TYPE));
+                    Op.Result queryable = block.op(invoke(md, query, quotedLambda));
 
-            return qp.createQuery(elementType, nextQueryExpression);
+                    block.op(_return(queryable));
+                } else {
+                    block.op(op);
+                }
+                return block;
+            });
+
+            return qp.createQuery(et, nextQueryExpression);
         }
 
-        @SuppressWarnings("unchecked")
-        default QueryResult<Stream<T>> elements() {
+        // Iterate
+        // Queryable -> Stream
+        default QueryResult elements() {
             TypeDesc resultType = type(type(Stream.class), elementType());
-            return (QueryResult<Stream<T>>) insertQueryResult(resultType, "elements");
+            return insertQueryResult("elements", resultType);
         }
 
-        @SuppressWarnings("unchecked")
-        default QueryResult<Long> count() {
-            return (QueryResult<Long>) insertQueryResult(TypeDesc.LONG, "count");
+        // Count
+        // Queryable -> Long
+        default QueryResult count() {
+            return insertQueryResult("count", TypeDesc.LONG);
         }
 
-        private QueryResult<?> insertQueryResult(TypeDesc resultType, String methodName) {
+        private QueryResult insertQueryResult(String name, TypeDesc resultType) {
             QueryProvider qp = provider();
 
-            // Copy function expression, replacing return operation
+            // Copy function expression, replacing return type
             FuncOp currentQueryExpression = expression();
-            TypeDesc queryResultType = TypeDesc.type(qp.queryResultType(), resultType);
-            FuncOp queryResultExpression = func("queryResult",
-                    methodType(queryResultType, currentQueryExpression.funcDescriptor().parameters()))
+            FuncOp nextQueryExpression = func("queryresult",
+                    methodType(qp.queryResultType(), currentQueryExpression.funcDescriptor().parameters()))
                     .body(b -> b.inline(currentQueryExpression, b.parameters(), (block, query) -> {
-                        MethodDesc md = method(qp.queryableType(), methodName, methodType(qp.queryResultType()));
-                        Op.Result queryResult = block.op(invoke(queryResultType, md, query));
+                        MethodDesc md = method(qp.queryableType(), name, methodType(qp.queryResultType()));
+                        Op.Result queryResult = block.op(invoke(md, query));
 
                         block.op(_return(queryResult));
                     }));
-            return qp.createQueryResult(resultType, queryResultExpression);
+            return qp.createQueryResult(resultType, nextQueryExpression);
         }
     }
 
-    public interface QueryResult<T> {
+    public interface QueryResult {
         TypeDesc resultType();
 
-        // Queryable<T> -> QueryResult<T>
+        // Queryable -> QueryResult
         FuncOp expression();
 
         Object execute();
@@ -141,28 +139,27 @@ public class TestLinqUsingQuotable {
 
         TypeDesc queryResultType();
 
-        Queryable<?> createQuery(TypeDesc elementType, FuncOp queryExpression);
+        Queryable createQuery(TypeDesc elementType, FuncOp expression);
 
-        QueryResult<?> createQueryResult(TypeDesc resultType, FuncOp expression);
+        QueryResult createQueryResult(TypeDesc resultType, FuncOp expression);
 
-        <T> Queryable<T> newQuery(Class<T> elementType);
+        Queryable newQuery(TypeDesc elementType);
     }
 
 
     // Query implementation
 
-    public static final class TestQueryable<T> implements Queryable<T> {
+    public static final class TestQueryable implements Queryable {
         final TypeDesc elementType;
         final TestQueryProvider provider;
         final FuncOp expression;
 
-        TestQueryable(Class<T> tableClass, TestQueryProvider qp) {
-            this.elementType = type(tableClass);
-            this.provider = qp;
+        TestQueryable(TypeDesc elementType, TestQueryProvider provider) {
+            this.elementType = elementType;
+            this.provider = provider;
 
-            TypeDesc queryableType = TypeDesc.type(qp.queryableType(), elementType);
             // Initial expression is an identity function
-            var funDescriptor = methodType(queryableType, queryableType);
+            var funDescriptor = methodType(provider().queryableType(), provider().queryableType());
             this.expression = func("query", funDescriptor)
                     .body(b -> b.op(_return(b.parameters().get(0))));
         }
@@ -189,7 +186,7 @@ public class TestLinqUsingQuotable {
         }
     }
 
-    public record TestQueryResult<T>(TypeDesc resultType, FuncOp expression) implements QueryResult<T> {
+    public record TestQueryResult(TypeDesc resultType, FuncOp expression) implements QueryResult {
         @Override
         public Object execute() {
             // @@@ Compile/translate the expression and execute it
@@ -217,18 +214,18 @@ public class TestLinqUsingQuotable {
         }
 
         @Override
-        public Queryable<?> createQuery(TypeDesc elementType, FuncOp expression) {
-            return new TestQueryable<>(elementType, this, expression);
+        public TestQueryable createQuery(TypeDesc elementType, FuncOp expression) {
+            return new TestQueryable(elementType, this, expression);
         }
 
         @Override
-        public QueryResult<?> createQueryResult(TypeDesc resultType, FuncOp expression) {
-            return new TestQueryResult<>(resultType, expression);
+        public QueryResult createQueryResult(TypeDesc resultType, FuncOp expression) {
+            return new TestQueryResult(resultType, expression);
         }
 
         @Override
-        public <T> Queryable<T> newQuery(Class<T> elementType) {
-            return new TestQueryable<>(elementType, this);
+        public Queryable newQuery(TypeDesc elementType) {
+            return new TestQueryable(elementType, this);
         }
     }
 
@@ -246,15 +243,16 @@ public class TestLinqUsingQuotable {
     public void testSimpleQuery() {
         QueryProvider qp = new TestQueryProvider();
 
-        QueryResult<Stream<String>> qr = qp.newQuery(Customer.class)
-                .where(c -> c.city.equals("London"))
-                .select(c -> c.contactName).elements();
+        QueryResult qr = qp.newQuery(type(Customer.class))
+                // c -> c.city.equals("London")
+                .where((Customer c) -> c.city.equals("London"))
+                // c -> c.contactName
+                .select((Customer c) -> c.contactName).elements();
 
         qr.expression().writeTo(System.out);
 
-        @SuppressWarnings("unchecked")
-        QueryResult<Stream<String>> qr2 = (QueryResult<Stream<String>>) Interpreter.invoke(MethodHandles.lookup(),
-                qr.expression(), qp.newQuery(Customer.class));
+        QueryResult qr2 = (QueryResult) Interpreter.invoke(MethodHandles.lookup(),
+                qr.expression(), qp.newQuery(type(Customer.class)));
 
         qr2.expression().writeTo(System.out);
