+++ test/jdk/java/lang/reflect/code/ad/TestForwardAutoDiff.java
@@ -40,6 +40,7 @@ import java.util.stream.Stream;
 
 /*
  * @test
+ * @enablePreview
  * @run testng TestForwardAutoDiff
  */
 
@@ -60,13 +61,13 @@ public class TestForwardAutoDiff {
         Block.Parameter x = f.body().entryBlock().parameters().get(0);
         Block.Parameter y = f.body().entryBlock().parameters().get(1);
 
-        CoreOps.FuncOp dff_dx = ExpressionElimination.eliminate(ForwardDifferentiation.diff(f, x));
+        CoreOps.FuncOp dff_dx = ExpressionElimination.eliminate(ForwardDifferentiation.partialDiff(f, x));
         dff_dx.writeTo(System.out);
         MethodHandle dff_dx_mh = generate(dff_dx);
         Assert.assertEquals((double) dff_dx_mh.invoke(0.0, 1.0), df_dx(0.0, 1.0));
         Assert.assertEquals((double) dff_dx_mh.invoke(PI_4, PI_4), df_dx(PI_4, PI_4));
 
-        CoreOps.FuncOp dff_dy = ExpressionElimination.eliminate(ForwardDifferentiation.diff(f, y));
+        CoreOps.FuncOp dff_dy = ExpressionElimination.eliminate(ForwardDifferentiation.partialDiff(f, y));
         dff_dy.writeTo(System.out);
         MethodHandle dff_dy_mh = generate(dff_dy);
         Assert.assertEquals((double) dff_dy_mh.invoke(0.0, 1.0), df_dy(0.0, 1.0));
@@ -110,7 +111,7 @@ public class TestForwardAutoDiff {
 
         Block.Parameter x = f.body().entryBlock().parameters().get(0);
 
-        CoreOps.FuncOp df_dx = ForwardDifferentiation.diff(f, x);
+        CoreOps.FuncOp df_dx = ForwardDifferentiation.partialDiff(f, x);
         df_dx.writeTo(System.out);
         MethodHandle df_dx_mh = generate(df_dx);
